Modified DoctorService, DoctorController, added @Json ignore to entities

+++ src/main/java/com/spring/hospital/app/services/DoctorService.java
@@ -1,6 +1,7 @@
 package com.spring.hospital.app.services;
 
 import com.spring.hospital.app.entities.Doctor;
+import com.spring.hospital.app.entities.Patient;
 import com.spring.hospital.app.entities.Treatment;
 import com.spring.hospital.app.repositories.DoctorRepos;
 import org.springframework.beans.factory.annotation.Autowired;
@@ -29,12 +30,18 @@ public class DoctorService {
         return doctor.orElse(null);
     }
 
+    public List <Treatment> getDoctorTreatmentsByDoctorId(Integer id){
+        Doctor doctor = doctorRepos.getById(id);
+        List<Treatment> treatments = doctor.getTreatments();
+        return treatments;
+    }
     public void saveDoctor(Doctor doctor){
         doctorRepos.save(doctor);
     }
 
     public void deleteDoctorById(Integer id){
-        doctorRepos.deleteById(id);
+        if (doctorRepos.existsById(id))
+            doctorRepos.deleteById(id);
     }
 
     public String getDoctorFirstMiddleLastNameStringById(Integer id){
